server/requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
Flask-Bcrypt>=1.0.0
python-dotenv>=0.19.0
Flask-CORS>=3.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
google-generativeai>=0.3.0
